Refact: Separated SkillAutocomplete
